requirements.txt
fastapi
uvicorn
PyMuPDF
google-generativeai
requests
python-multipart